comments/pyproject.toml
[tool.poetry]
name = "comments "
version = "0.1.0"
description = ""
authors = ["zifrit <zifrit@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
uvicorn = "^0.32.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
pydantic-settings = "^2.7.0"
fastapi = "^0.115.6"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
